Modularized the homepage.js
