+++ src/modules/homepage.js
@@ -1,8 +1,5 @@
 import heart from '../assets/fullheart.svg';
 import notLiked from '../assets/emptyheart.svg';
-import addLike from './addLike.js';
-import endpoint, { ArtPiecesLikesData } from './api.js';
-import getLikes from './getLikes.js';
 
 const emptyheart = new Image();
 emptyheart.src = heart;
@@ -26,31 +23,6 @@ const Homepage = async (i, artPieces) => {
       </div>
     `;
   }
-};
-
-export const displayLikes = async (endpoint) => {
-  const data = await getLikes(endpoint.likes);
-  ArtPiecesLikesData.push(...data);
-  ArtPiecesLikesData.forEach((likeData) => {
-    const likesCounter = document.querySelector(`#${likeData.item_id}-like`);
-    if (likesCounter) {
-      likesCounter.innerText = `${likeData.likes} likes`;
-    }
-  });
-};
-
-const handleLike = async (event) => {
-  const heartId = event.target;
-  const artId = heartId.id;
-  await addLike(endpoint.likes, artId);
-  displayLikes(endpoint.likes);
-};
-
-export const createLikes = async () => {
-  const likeButtons = document.querySelectorAll('.likes-count-icon');
-  likeButtons.forEach((likeButton) => {
-    likeButton.addEventListener('click', handleLike);
-  });
 };
 
 export default Homepage;
